+++ commands/test.ps1
@@ -1,0 +1,2 @@
+$exitCode = Invoke-DevToolsTestSuite
+exit $exitCode